Python projects
This python programs only work with
 Windows, not macOS. Some stuff might be in portuguese, but if you understand english you can make yourself through.
 Get_internet_password searches for the names of the wifi connections and their passwords that are already in your computer and stores both in the Lista_Wifi.txt document. "Lista_Wifi_Secure.txt" is the back up of the last version of Lista_Wifi before the iteration, so in case any of some problem you wont loose everything. I like it because it saves paper and it also saves you the time for the person to give you their wifi password, the only downgrade beeing that you need to have Python installed in that PC, or you can convert it to an .exe, do as you will.
*The code from Wifi_password.py and funcoes.py is heavily inspired on the tutorial of NeuralLine, a youtube channel that is focused on code and coding* please, check him up, he's very good.

The Morse_translator is only for windows but if delete the part that makes the sound it works for Mac. This program was made intirely by me, with the help of the community. It translates what you type for it to translation. It can even save the translation on a .txt file.

Please don't use these programs in a malicious way, they are ment for personal, moral and legal use. Thank you for your comprehension.

Enjoy