Conway's Game of Life (Uni) and Albums_manager
First of all, I want to say sorry because of my english errors, speacially those that are on the first post, anyway....

Conway's Game of Life: This is a uni work, so it's 99% written in portuguese, and it wasn't coded 100% by me. this is a Conway's game of life attempt. If you wan't to try it you need to run "testa_anima_mundo().py". After running you need to write a name of one of those text files, write the number of iterations, aka frames that you'll watch and by last you need to say in seconds what is the delay you desire within each second. Enjoy it took a bit of work to code.

Albums Manager: This one you anglo-saxons folks will understand. It's an album manager, this one is very personal, btw I recommend you to check the albums in the .txt files, they are very good, some of my favorite, actually.  This program is a list manager, like a wish list manager and it also keeps the albums that you have. The info is formated in specific way  and it's presented to you a proper list, I feel like this could be a base for a phone app or something in that sense, all things that I don't know how to make them, yet*.

Again, sry for any typos and one last thing:
I don't mind for people to use my work for free, you can copy and paste whatever you want, that is why this repository is public, right? But I just ask you to mention me, MildDrake and this repositorie, https://github.com/MildDrake/simple-py.git  in your  work in any way. Through #comments# or prints, print("This code was made by, MildDrake"), whatever, you got the point.

Thank you, everybody